fastapi/app/requirements.txt
fastapi
uvicorn
kafka-python
psycopg2-binary
clickhouse-driver
requests
pydantic